conversation_feeder/requirements.txt
fastapi
uvicorn
pydantic-settings
asyncpg
SQLAlchemy
pgvector-python
google-generativeai
python-multipart
Jinja2

# Testing & Linting
pytest
pytest-asyncio
pytest-mock
httpx
ruff